Refactor skill extraction into modular analyzers
Split skill extraction logic into separate modules: skill_code_stats, skill_models, skill_patterns, skill_proficiency, and skill_scanners. SkillExtractor now orchestrates evidence gathering, code stats collection, and proficiency scoring using these components. Improves maintainability and extensibility by isolating data models, regex patterns, and proficiency heuristics.
